non-contiguous checks at beginning of functions
This pr adds a contiguous checks at the beginning of most functions. Most of our
functions don't support non-contiguous tensors.

Also changed the definition of IsContiguous a bit. Noticed that in
pytorh a contiguous tensor must have its last dimension's stride set to
1 AND not have other strides set to zero. If this doesn't work let me
know!

Another thing that is needed in the future is a function to turn a
non-contiguous tensor into a contiguous one. Not really sure how to
implement this right now.

Check out this discussion for more: https://github.com/OpenMined/OpenMined/issues/205