new isContiguous algorithm that checks all cases
Old isContiguous wasn't checking all cases of contiguity, borrowed an
algorithm from here https://github.com/torch/torch7/blob/master/lib/TH/generic/THTensor.c#L643

This algorithm basically just checks that the strides are still a valid
output from the algorithm that creats the strides. If not then it is
known to be not contiguous.

Also modified the test to pass. It was using transpose before to check
non-contiguity but our implementation of transpose leaves the tensor in
a contiguous state so the test was failing. The expand function makes
the tensor non-contiguous.